Prefer file info config and fix trailing / in directory name bug
Previously, the plugin only used HWM/LWM config, a la:
`pfc.diskusage 20m 30m purgeinterval 60s`
where 30m/20m are used as the "start purging" & "stop purging" setpoints.

Now, it prefers file info, a la:
`pfc.diskusage 1g 2g files 40m 45m 50m purgeinterval 60s`
where the 50m/40m are used as the setpoints.

The benefit of this approach is that the file info numbers _only_ count
files that are cached, rather than the entire disk usage which is polluted
by any other usage of the disk that's not part of the cache.

This also fixes a (new?) bug where member directories in the `m_list` need
a trailing / to be constructed correctly when passed to XRootD but didn't
have one.